requirements.txt
streamlit
pandas
matplotlib
pg8000
python-dotenv
tableauserverclient
tableau-api-lib
boto3